improve version comprehesion
Uses `shutil.which()` to get the path to an existing executable. If specified version is a semantic version number, then it is used as a path to the executable.

Now throws an assertion error if tool executable could not be discovered.

Moved tool skipping to `capture_clang_tools_output()` and show executable's version in logs (for transparency).

added unit test to cover changes